Create IV_graph.ipynb
Here is the code behind the IV curve plotting.
I used the data from Wed. that Nick and I recorded.
There was some cloud coverage that affected the numbers, but I wanted to use this as practice for graphing an IV curve. 

The link for the visual view of the graph is found within the README.md that is accompanied with this file